Appended Warnings Switch
Appended new feature: now you can set warnings for the compiler to show. Configure warnings in the parameters.dat file.

Appended behavior for the 1800COMPILE.bat file to allow for compilation of other scripts rather than having to return to the main window.

Reworked a few lines and replaced `::` comments with `@REM` comments to prevent strange and obtuse batch errors.